app/pyproject.toml
[project]
name = "multi-agent-ai-bot"
version = "0.1.0"
description = "LangGraph multi-agent chatbot (MCP, RAG, memory)."
authors = [{name = "Your Name", email = "you@example.com"}]
readme = "README.md"
requires-python = ">=3.11"

# ───────────────────────── Runtime dependencies ─────────────────────────
dependencies = [
  # LangChain core + provider helpers
  "langchain>=0.2.1",
  "langchain-community>=0.2.1",
  "langchain-aws>=0.1.7",
  "langchain-openai>=0.1.0",

  # Vector store & cache
  "weaviate-client>=3.26.7,<4.0.0",   # v3 pinned (SDK-v4 breaking changes)
  "redis>=5.0.0",

  # Web stack
  "fastapi>=0.110.0",
  "uvicorn[standard]>=0.27.0",

  # Config / HTTP / CLI
  "python-dotenv>=1.0.1",
  "pydantic>=2,<3",
  "pydantic-settings>=2.1",
  "httpx>=0.27.0",
  "typer[all]>=0.9.0",

  # Tool-specific deps
  "kubernetes>=26.1.0",
  "isodate>=0.6.1"
]

# ───────────────────────── Dev / CI extras ─────────────────────────
[project.optional-dependencies]
dev = [
  "pytest>=8.1.0",
  "ruff>=0.4.3",
  "mypy>=1.10.0"
]

# ───────────────────────── Build backend ─────────────────────────
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

# ───────────────────────── Package discovery ─────────────────────────
[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
